requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Deprecated==1.3.1
ecdsa==0.19.1
fastapi==0.128.0
fonttools==4.61.1
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.47.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
networkx==3.2.1
numpy==2.4.1
packaging==25.0
passlib==1.7.4
pillow==12.1.0
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.1.1
PyJWT==2.10.1
PyNaCl==1.6.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.65.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wheel==0.45.1
wrapt==2.0.1
